Add new marriage construct for code climate
